examples/python/beeai-handoff/pyproject.toml
[project]
name = "beeai-handoff"
version = "0.1.0"
description = "Demonstration of handoff pattern"
requires-python = ">=3.11, <4.0"
dependencies = [
    "acp-sdk",
    "beeai-framework>=0.1.17",
    "pydantic-settings>=2.9.0",
]

[tool.uv.sources]
acp-sdk = { workspace = true }
